created a useLaunchService hook
