+++ example/src/features/calculator/CalculatorForm.tsx
@@ -1,11 +1,13 @@
 import React, { useEffect } from "react";
-import { useCalculator } from "./calculator.hooks";
+import { useLaunchService } from "launch.io";
 
 const CalculatorForm = () => {
-  let { value, increase, decrease } = useCalculator();
+  let { value, increase, decrease } = useLaunchService("calculator");
+
   useEffect(() => {
     console.log("CalculatorForm -> RENDER", value);
   });
+
   return (
     <div>
       <p>Value: {value}</p>
