+++ src/index.ts
@@ -1,4 +1,2 @@
-import { useLaunchProvider } from "./components/LaunchProvider";
-import useLaunch from "./hooks/useLaunch";
-import { useLaunchSelector } from "./hooks/useLaunchSelector";
-export { useLaunch, useLaunchSelector, useLaunchProvider };
+export * from "./components/LaunchProvider";
+export * from "./hooks/useLaunch";
